Revert moving CardType
