+++ farebot-card-cepas/src/main/java/com/codebutler/farebot/card/cepas/CEPASCard.java
@@ -26,13 +26,11 @@ package com.codebutler.farebot.card.cepas;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-
-import com.codebutler.farebot.card.Card;
-import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.base.ui.FareBotUiTree;
 import com.codebutler.farebot.base.util.ByteArray;
+import com.codebutler.farebot.card.Card;
+import com.codebutler.farebot.card.CardType;
 import com.google.auto.value.AutoValue;
-
 import java.text.DateFormat;
 import java.text.NumberFormat;
 import java.util.Date;
