+++ farebot-card-classic/src/main/java/com/codebutler/farebot/card/classic/raw/RawClassicCard.java
@@ -24,9 +24,8 @@
 package com.codebutler.farebot.card.classic.raw;
 
 import android.support.annotation.NonNull;
-
 import com.codebutler.farebot.base.util.ByteArray;
-import com.codebutler.farebot.transit.registry.annotations.CardType;
+import com.codebutler.farebot.card.CardType;
 import com.codebutler.farebot.card.RawCard;
 import com.codebutler.farebot.card.classic.ClassicCard;
 import com.codebutler.farebot.card.classic.ClassicSector;
@@ -36,7 +35,6 @@ import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
-
 import java.util.Date;
 import java.util.List;
