+++ farebot-card-ultralight/src/main/java/com/codebutler/farebot/card/ultralight/UltralightCard.java
@@ -25,13 +25,11 @@ package com.codebutler.farebot.card.ultralight;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
-
-import com.codebutler.farebot.card.Card;
-import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.base.ui.FareBotUiTree;
 import com.codebutler.farebot.base.util.ByteArray;
+import com.codebutler.farebot.card.Card;
+import com.codebutler.farebot.card.CardType;
 import com.google.auto.value.AutoValue;
-
 import java.util.Date;
 import java.util.List;
