+++ farebot-card-ultralight/src/main/java/com/codebutler/farebot/card/ultralight/raw/RawUltralightCard.java
@@ -23,16 +23,14 @@
 package com.codebutler.farebot.card.ultralight.raw;
 
 import android.support.annotation.NonNull;
-
 import com.codebutler.farebot.base.util.ByteArray;
-import com.codebutler.farebot.transit.registry.annotations.CardType;
+import com.codebutler.farebot.card.CardType;
 import com.codebutler.farebot.card.RawCard;
 import com.codebutler.farebot.card.ultralight.UltralightCard;
 import com.codebutler.farebot.card.ultralight.UltralightPage;
 import com.google.auto.value.AutoValue;
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
-
 import java.util.Date;
 import java.util.List;
