+++ farebot-transit-myki/src/main/java/com/codebutler/farebot/transit/myki/MykiTransitFactory.java
@@ -21,14 +21,15 @@
 package com.codebutler.farebot.transit.myki;
 
 import android.support.annotation.NonNull;
-
 import com.codebutler.farebot.base.util.ByteUtils;
 import com.codebutler.farebot.base.util.Luhn;
 import com.codebutler.farebot.card.desfire.DesfireCard;
 import com.codebutler.farebot.card.desfire.StandardDesfireFile;
 import com.codebutler.farebot.transit.TransitFactory;
 import com.codebutler.farebot.transit.TransitIdentity;
+import com.codebutler.farebot.transit.registry.annotations.TransitCard;
 
+@TransitCard
 public class MykiTransitFactory implements TransitFactory<DesfireCard, MykiTransitInfo> {
 
     @Override
