+++ farebot-transit-registry-annotations/src/main/java/com/codebutler/farebot/transit/registry/annotations/TransitCard.java
@@ -11,5 +11,4 @@ import static java.lang.annotation.RetentionPolicy.CLASS;
 @Retention(CLASS)
 @CrumbProducer
 public @interface TransitCard {
-  CardType value();
 }
